Fixing compilation. TODO: VERIFICAR

+++ src/C++/Message.cpp
@@ -178,7 +178,7 @@ std::string& Message::toString( std::string& str,
                                 int checkSumField ) const
 {
   size_t length = bodyLength( beginStringField, bodyLengthField, checkSumField );
-  m_header.setField( IntField(bodyLengthField, length) );
+  m_header.setField( IntField(bodyLengthField, static_cast<int>(length)) );
   m_trailer.setField( CheckSumField(checkSumField, checkSum(checkSumField)) );
 
 #if defined(_MSC_VER) && _MSC_VER < 1300
